TelegramGetAuthToken/requirements.txt
Telethon==1.38.1
uvicorn==0.34.0
fastapi==0.115.8
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
pydantic_core==2.27.2